+++ src/main/java/org/opensearch/knn/index/codec/nativeindex/remote/RemoteIndexBuildStrategy.java
@@ -146,20 +146,20 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
 
             // 4. Download index file and write to indexOutput
             readFromRepository(indexInfo, repositoryContext, remoteBuildStatusResponse);
-
             success = true;
+            return;
         } catch (Exception e) {
-            fallbackStrategy.buildAndWriteIndex(indexInfo);
+            log.error("Failed to build index remotely: " + indexInfo, e);
         } finally {
             metrics.endRemoteIndexBuildMetrics(success);
         }
+        fallbackStrategy.buildAndWriteIndex(indexInfo);
     }
 
     /**
      * Writes the required vector and doc ID data to the repository
      */
-    private void writeToRepository(RepositoryContext repositoryContext, BuildIndexParams indexInfo) throws IOException,
-        InterruptedException {
+    private void writeToRepository(RepositoryContext repositoryContext, BuildIndexParams indexInfo) {
         VectorRepositoryAccessor vectorRepositoryAccessor = repositoryContext.vectorRepositoryAccessor;
         boolean success = false;
         metrics.startRepositoryWriteMetrics();
@@ -172,8 +172,7 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
             );
             success = true;
         } catch (InterruptedException | IOException e) {
-            log.debug("Repository write failed for vector field [{}]", indexInfo.getFieldName());
-            throw e;
+            throw new RuntimeException(String.format("Repository write failed for vector field [%s]", indexInfo.getFieldName()), e);
         } finally {
             metrics.endRepositoryWriteMetrics(success);
         }
@@ -183,8 +182,7 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
      * Submits a remote build request to the remote index build service
      * @return RemoteBuildResponse containing the response from the remote service
      */
-    private RemoteBuildResponse submitBuild(RepositoryContext repositoryContext, BuildIndexParams indexInfo, RemoteIndexClient client)
-        throws IOException {
+    private RemoteBuildResponse submitBuild(RepositoryContext repositoryContext, BuildIndexParams indexInfo, RemoteIndexClient client) {
         final RemoteBuildResponse remoteBuildResponse;
         boolean success = false;
         metrics.startBuildRequestMetrics();
@@ -200,8 +198,7 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
             success = true;
             return remoteBuildResponse;
         } catch (IOException e) {
-            log.debug("Submit vector build failed for vector field [{}]", indexInfo.getFieldName());
-            throw e;
+            throw new RuntimeException(String.format("Submit vector build failed for vector field [%s]", indexInfo.getFieldName()), e);
         } finally {
             metrics.endBuildRequestMetrics(success);
         }
@@ -216,7 +213,7 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
         RemoteBuildResponse remoteBuildResponse,
         BuildIndexParams indexInfo,
         RemoteIndexClient client
-    ) throws IOException, InterruptedException {
+    ) {
         RemoteBuildStatusResponse remoteBuildStatusResponse;
         metrics.startWaitingMetrics();
         try {
@@ -225,11 +222,11 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
                 .build();
             RemoteIndexWaiter waiter = RemoteIndexWaiterFactory.getRemoteIndexWaiter(client);
             remoteBuildStatusResponse = waiter.awaitVectorBuild(remoteBuildStatusRequest);
-            metrics.endWaitingMetrics();
             return remoteBuildStatusResponse;
         } catch (InterruptedException | IOException e) {
-            log.debug("Await vector build failed for vector field [{}]", indexInfo.getFieldName(), e);
-            throw e;
+            throw new RuntimeException(String.format("Await index build failed for vector field [%s]", indexInfo.getFieldName()), e);
+        } finally {
+            metrics.endWaitingMetrics();
         }
     }
 
@@ -240,7 +237,7 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
         BuildIndexParams indexInfo,
         RepositoryContext repositoryContext,
         RemoteBuildStatusResponse remoteBuildStatusResponse
-    ) throws IOException {
+    ) {
         metrics.startRepositoryReadMetrics();
         boolean success = false;
         try {
@@ -250,8 +247,7 @@ public class RemoteIndexBuildStrategy implements NativeIndexBuildStrategy {
             );
             success = true;
         } catch (Exception e) {
-            log.debug("Repository read failed for vector field [{}]", indexInfo.getFieldName());
-            throw e;
+            throw new RuntimeException(String.format("Repository read failed for vector field [%s]", indexInfo.getFieldName()), e);
         } finally {
             metrics.endRepositoryReadMetrics(success);
         }